add useDeployedContractInfo hook and remove getDeployedContract util
